Starting to refactor tree query to separate classes
